Implementation and bug fix of the Naive Engine.
Fix of issues with the function parameter for evaluation.
	Change to a binary function for propagation.
	Fix of the min Function.

Addition of Least and Greatest element to the Engine Base. This is used
to keep the axiom for the parameter function when taken on empty sets.

Fix and improvement to the unification algorithm
	Now uses a single UnionFind Unifier per rule, this is more efficient
and works for a larger set of rules.

Each iteration now also recomputes the ground facts as per the naive
algorithm specification.

Various bug fix to the naive engine, it now works without error.
	TODO validation of result