Addition of a lock to the FactsRepository as a defensive measure.
This is added in preparation for concurrently processing rules during an
iteration. The facts Repository should never be updated during an
iteration, only read. While it should not happen that it is modified
according to the inference algorithm, this is preventive to quickly fail
if an error is introduced in the implementation.