Implementation of the Semi-Naive CandidateTupleGenerator.
This generator uses an helping data structure of annotated ordered list.
This lists separates a single list into classes of components to
distinguish the new facts from the old ones. Therefore the Generator is
able to easily iterate to always have at least a single fact that is new
in its tuple.